requirements.txt
pyserial==3.5
fastapi
fastapi[standard]
uvicorn
gunicorn